chore: fix props exported from a shared types file
